Add files via upload
Actually I recently made this type of script to check for my 3 Semester result. I used urllib2 and requests module to scrap through the result website.  This script does very simple thing it finds for the keyword of Result of Semester 3 if it founds nothing it goes into sleep for next 15 minutes and if it gets it sends notification to my Desktop ( I use ubuntu so notification won't work in windows ).